RSSEC-171 refactor: rename files
