+++ src/core/interfaces/accessTokenResponse.ts
