+++ src/core/interfaces/customerSignInResponse.ts
