+++ src/core/interfaces/loginRequest.ts
