+++ src/server/categories/getCategoryById.ts
