+++ src/server/categories/queryAllCategories.ts
