+++ src/server/customerLogin.ts
@@ -3,8 +3,8 @@ import App, { PageId } from '../pages/app/app';
 import { constants } from '../data/constants';
 import { showToast } from '../pages/logReg/utils/funcToastify.utils';
 import { ICustomerResponse } from '../core/interfaces/customerResponse';
-import { ILoginRequest } from '../core/interfaces/ginRequest';
-import { ICustomerSignInResponse } from '../core/interfaces/stomerSignInResponse';
+import { ILoginRequest } from '../core/interfaces/loginRequest';
+import { ICustomerSignInResponse } from '../core/interfaces/customerSignInResponse';
 
 export let dataCustomer: ICustomerResponse;
