+++ src/server/customerRegistration.ts
