+++ src/server/function/structureCategories.ts
@@ -4,10 +4,10 @@ import {
     IProductResp,
     ISubCategoryResp,
 } from '../../pages/catalog/interfaces/categoryResponse/categoryResponseInterface';
-import { QueryAllCategories } from '../categories/eryAllCategories';
+import { QueryAllCategories } from '../categories/queryAllCategories';
 // import { AllProductDetailsGQL } from '../products/AllProductDetailsGQL';
 // import { QueryAllProducts } from '../products/QueryAllProducts';
-import { QueryProductProjections } from '../products/ueryProductProjections';
+import { QueryProductProjections } from '../products/queryProductProjections';
 import { ICategoryResponse } from './interfaces';
 
 export let categoryResponse: ICategoryResp[] = [];
