+++ src/server/products/allProductDetailsGQL.ts
